+++ client/src/components/App.js
@@ -2,6 +2,7 @@ import React from 'react'
 import {Route, Switch, Redirect} from 'react-router-dom'
 import Landing from './Landing'
 import Main from './Main'
+import './App.css'
 
 const App = () => (
   <Switch>
