+++ client/src/components/Main.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import './variables.scss';
 
 .application-container {
   margin-top: $navbar-height;
