+++ client/src/components/Notices.js
@@ -1,8 +1,4 @@
 import React from 'react'
-import {compose} from 'redux'
-import {connect} from 'react-redux'
-import {withRefetch} from 'data-provider'
-import {noticesProvider} from '../dataProviders/noticesDataProviders'
 
 const Notices = ({dispatch}) => <div>Obstaravania placeholder</div>
