+++ client/src/components/Verejne/GoogleMap/index.js
@@ -4,7 +4,7 @@ import {map} from 'lodash'
 import Marker from './Marker'
 import classnames from 'classnames'
 import {connect} from 'react-redux'
-import {mapOptionsSelector, mapReferenceSelector, entitiesSelector} from '../../../state/verejneState'
+import {mapOptionsSelector, mapReferenceSelector, entitiesSelector} from '../../../selectors'
 import {initializeGoogleMap, fetchEntities, updateMapOptions} from '../../../actions/verejneActions'
 import {GOOGLE_MAP_API_KEY, createMapOptions, clusterOptions} from '../../../constants'
 import supercluster from 'points-cluster'
