+++ client/src/index.js
@@ -11,8 +11,6 @@ import App from './components/App'
 import getConfiguredStore from './configureStore'
 import {Provider} from 'react-redux'
 
-import './custom.css'
-
 window.Promise = Promise
 
 // a short-term fix for data-provider, should get fixed in next release
