+++ client/src/rootReducer.js
@@ -1,9 +1,9 @@
 // @flow
 import {forwardReducerTo} from './utils'
-import getInitialState from './initialState'
+import getInitialState from './state'
 
 import type {GenericAction} from './types/reduxTypes'
-import type {State} from './initialState'
+import type {State} from './state'
 
 const rootReducer = (state: State = getInitialState(), action: GenericAction<*, *>) => {
   const {reducer, path, payload} = action
